backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
deepface==0.0.79
tensorflow==2.15.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Face recognition dependencies
opencv-contrib-python==4.8.1.78
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2